# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "Body_upload_api_uploads_post": {
        "properties": {
          "file": {
            "contentMediaType": "application/octet-stream",
            "title": "File",
            "type": "string"
          }
        },
        "required": [
          "file"
        ],
        "title": "Body_upload_api_uploads_post",
        "type": "object"
      },
...
      "CatalogSource": {
        "properties": {
          "data_type": {
            "enum": [
              "vector",
              "raster"
            ],
            "title": "Data Type",
            "type": "string"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "id": {
            "title": "Id",
            "type": "string"
          },
          "kind": {
            "enum": [
              "osm_preset",
              "copernicus",
              "upload",
              "sample"
            ],
            "title": "Kind",
            "type": "string"
          },
          "label": {
            "title": "Label",
            "type": "string"
          },
          "layer_template": {
            "additionalProperties": true,
            "title": "Layer Template",
            "type": "object"
          }
        },
        "required": [
          "id",
          "kind",
          "label",
          "data_type",
          "layer_template"
        ],
        "title": "CatalogSource",
        "type": "object"
      },
...
      "ErrorResponse": {
        "description": "Fehlerantwort (``HTTPException``): ``detail`` ist die Meldung.",
        "properties": {
          "detail": {
            "title": "Detail",
            "type": "string"
          }
        },
        "required": [
          "detail"
        ],
        "title": "ErrorResponse",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "UploadResponse": {
        "properties": {
          "source": {
            "$ref": "#/components/schemas/CatalogSource"
          }
        },
        "required": [
          "source"
        ],
        "title": "UploadResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "description": "Token aus POST /api/auth/login als 'Authorization: Bearer <token>'. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "scheme": "bearer",
        "type": "http"
      },
      "QueryToken": {
        "description": "Dasselbe Token als Query-Parameter - nur für Ereignisstrom, Rasterbild und Download, die der Browser ohne Authorization-Header abruft. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft.",
        "in": "query",
        "name": "token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "description": "Backend-API zur interaktiven Demonstration des GeoFACT-Frameworks.",
    "title": "GeoFACT Web API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/uploads": {
      "post": {
        "description": "Speichert eine Vektor-, Raster- oder Tabellendatei und liefert das Layer-Fragment für die Konfiguration. Die Datei gehört dem anfragenden Nutzer.",
        "operationId": "upload_api_uploads_post",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_api_uploads_post"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UploadResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "400": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Dateiendung wird nicht unterstützt."
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "413": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Datei größer als die Obergrenze des Servers."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Datei hochladen",
        "tags": [
          "uploads"
        ]
      }
    },
    "/api/uploads/{source_id}": {
      "delete": {
        "description": "Löscht einen eigenen Upload endgültig.",
        "operationId": "delete_api_uploads__source_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "source_id",
            "required": true,
            "schema": {
              "title": "Source Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Token fehlt, ist ungültig oder abgelaufen. Nur wirksam, wenn GEOFACT_WEB_USERS gesetzt ist (FA24); ohne konfigurierte Nutzer wird kein Token geprüft."
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Upload unbekannt oder gehört einem anderen Nutzer."
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Upload löschen",
        "tags": [
          "uploads"
        ]
      }
    }
  },
...
}